Added circle, and quadratic functionality
Circle and quadratic functions work as long as the graph dimensions are square/equal (appears to be the case for linear as well, as playing around with the different min and max values eventually skews the line position), so I'll work on fixing that. Also need to add input verification and bounds checking for my two functions but that should be pretty easy so saving that for last.